Ajout des fichiers .js du dossier dist

+++ generateFights.js
@@ -176,11 +176,13 @@ simulFights({
 var setInt
 var fightWorker
 	clearInterval(setInt);
-	var urrl = window.location.href
+	var urrl
+
+if(typeof(window)!="undefined"){	urrl= window.location.href;
 	setInt = setInterval(function(){if(fightWorker)fightWorker.postMessage(5);
 	if(window.location.href!=urrl){urrl=window.location.href;	stopLoading();
 	if(fightWorker)fightWorker.terminate()}
-	},333)
+},333)}
 async function simulFights(arg){
 	
 	fetch(BRUTALISATOR+"generateFights.js")
